Update ui.R
Only Contig1 - Contig1001 are displayed in the side panel initially. However, the entire list of contigs can be searched. The user can input a letter (i.e. "G"), and a contig that was previously not displayed on the list will be displayed. The user can select such contig and proceed to happily select additional contigs as desired.